feat: add alert on deleting

+++ lab_5/components/Folder.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'expo-router';
-import { Text, TouchableOpacity, View } from 'react-native';
+import { Alert, Text, TouchableOpacity, View } from 'react-native';
 import * as FileSystem from 'expo-file-system';
 import FontAwesome from '@expo/vector-icons/FontAwesome';
 import { rootPath } from '~/const';
@@ -14,8 +14,22 @@ export default function Folder({
   refresh: () => void;
 }) {
   const handleDelete = async () => {
-    await FileSystem.deleteAsync(`${rootPath}${fullPath}`, { idempotent: true });
-    refresh();
+    Alert.alert(
+      `Delete folder ${name}`,
+      undefined,
+      [
+        { text: 'Cancel', style: 'cancel' },
+        {
+          text: 'Delete',
+          style: 'destructive',
+          onPress: async () => {
+            await FileSystem.deleteAsync(`${rootPath}${fullPath}`, { idempotent: true });
+            refresh();
+          },
+        },
+      ],
+      { cancelable: true }
+    );
   };
 
   return (
